+++ src/jmcomic/jm_config.py
@@ -1,3 +1,8 @@
+import os
+import sys
+
+sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+
 from common import time_stamp, str_to_list, field_cache, ProxyBuilder
 from loguru import logger