+++ src/jmcomic/jm_plugin.py
@@ -1,6 +1,11 @@
 """
 该文件存放的是option插件
 """
+import os
+import sys
+
+sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+
 import os.path
 
 from .jm_option import *