+++ src/jmcomic/jm_toolkit.py
@@ -1,3 +1,8 @@
+import os
+import sys
+
+sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+
 from PIL import Image
 
 from .jm_exception import *