+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### 2024/7/3
 - add close main dialog ask quit
 - disabled retry on download image fail 
+- already download image process
 
 ## v1.1.8-beta.2.240628
 ### 2024/6/28